+++ tech1-framework-b2b-base-security-jwt/src/main/java/io/tech1/framework/b2b/base/security/jwt/tests/random/BaseSecurityJwtRandomUtility.java
@@ -18,18 +18,17 @@ import io.tech1.framework.domain.http.requests.UserRequestMetadata;
 import io.tech1.framework.domain.properties.base.TimeAmount;
 import io.tech1.framework.domain.system.reset_server.ResetServerStatus;
 import lombok.experimental.UtilityClass;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 
 import java.time.LocalDateTime;
 import java.time.temporal.ChronoUnit;
 import java.util.Date;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static io.tech1.framework.b2b.base.security.jwt.domain.jwt.JwtTokenValidatedClaims.getIssuedAt;
+import static io.tech1.framework.b2b.base.security.jwt.utilities.SpringAuthoritiesUtility.getSimpleGrantedAuthorities;
 import static io.tech1.framework.domain.base.AbstractAuthority.SUPER_ADMIN;
 import static io.tech1.framework.domain.utilities.random.RandomUtility.*;
 import static io.tech1.framework.domain.utilities.time.DateUtility.convertLocalDateTime;
@@ -45,15 +44,11 @@ public class BaseSecurityJwtRandomUtility {
                 Username.random(),
                 Password.random(),
                 randomZoneId(),
-                authorities(SUPER_ADMIN),
+                getSimpleGrantedAuthorities(SUPER_ADMIN),
                 Email.random(),
                 randomString(),
                 new HashMap<>()
         );
-    }
-
-    public static List<SimpleGrantedAuthority> authorities(String... authorities) {
-        return Stream.of(authorities).map(SimpleGrantedAuthority::new).toList();
     }
 
     public static Set<JwtAccessToken> accessTokens(String... accessTokens) {
@@ -67,7 +62,7 @@ public class BaseSecurityJwtRandomUtility {
         var expiration = convertLocalDateTime(LocalDateTime.now(UTC).plus(timeAmount.getAmount(), timeAmount.getUnit()), UTC);
         claims.setIssuedAt(getIssuedAt());
         claims.setExpiration(expiration);
-        claims.put("authorities", authorities("admin", "user"));
+        claims.put("authorities", getSimpleGrantedAuthorities("admin", "user"));
         return claims;
     }
 
@@ -79,12 +74,12 @@ public class BaseSecurityJwtRandomUtility {
         var expiration = new Date(currentTimestamp - 1000);
         claims.setIssuedAt(issuedAt);
         claims.setExpiration(expiration);
-        claims.put("authorities", authorities("admin", "user"));
+        claims.put("authorities", getSimpleGrantedAuthorities("admin", "user"));
         return claims;
     }
 
     public static InvitationCode randomInvitationCode() {
-        return new InvitationCode(InvitationCodeId.random(), Username.random(), authorities(SUPER_ADMIN), randomString(), Username.random());
+        return new InvitationCode(InvitationCodeId.random(), Username.random(), getSimpleGrantedAuthorities(SUPER_ADMIN), randomString(), Username.random());
     }
 
     public static UserSession randomPersistedSession() {