+++ tech1-framework-b2b-base-security-jwt/src/test/java/io/tech1/framework/b2b/base/security/jwt/utils/impl/SecurityJwtTokenUtilsImplTest.java
@@ -31,7 +31,7 @@ import java.time.ZoneId;
 import java.util.List;
 import java.util.stream.Stream;
 
-import static io.tech1.framework.b2b.base.security.jwt.tests.random.BaseSecurityJwtRandomUtility.authorities;
+import static io.tech1.framework.b2b.base.security.jwt.utilities.SpringAuthoritiesUtility.getSimpleGrantedAuthorities;
 import static io.tech1.framework.domain.tests.constants.TestsDTFsConstants.DEFAULT_DATE_FORMAT_PATTERN;
 import static io.tech1.framework.domain.tests.constants.TestsZoneIdsConstants.EET_TIME_ZONE;
 import static io.tech1.framework.domain.utilities.random.RandomUtility.randomZoneId;
@@ -133,7 +133,7 @@ class SecurityJwtTokenUtilsImplTest {
     void createJwtAccessTokenTest() {
         // Arrange
         var username = Username.of("multiuser43");
-        var authorities = authorities("admin", "user");
+        var authorities = getSimpleGrantedAuthorities("admin", "user");
         var creationParams = new JwtTokenCreationParams(username, authorities, randomZoneId());
 
         // Act
@@ -155,7 +155,7 @@ class SecurityJwtTokenUtilsImplTest {
     void createJwtRefreshTokenTest() {
         // Arrange
         var expectedUsername = Username.of("multiuser43");
-        var authorities = authorities("admin", "user");
+        var authorities = getSimpleGrantedAuthorities("admin", "user");
         var creationParams = new JwtTokenCreationParams(expectedUsername, authorities, randomZoneId());
 
         // Act
@@ -177,7 +177,7 @@ class SecurityJwtTokenUtilsImplTest {
     @MethodSource("createJwtTokenTest")
     void createJwtTokenTest(TimeAmount timeAmount) {
         // Arrange
-        var creationParams = new JwtTokenCreationParams(Username.testsHardcoded(), authorities("user"), randomZoneId());
+        var creationParams = new JwtTokenCreationParams(Username.testsHardcoded(), getSimpleGrantedAuthorities("user"), randomZoneId());
 
         // Act
         var jwtToken = this.componentUnderTest.createJwtToken(creationParams, timeAmount);